fix: Switch로 파일 변경 및 size props 삭제
