+++ docs/formInputs/Switch.mdx
@@ -1,30 +1,29 @@
 ---
 menu: Form Inputs
-name: SwitchButton
-route: /form-inputs/switch-button
+name: Switch
+route: /form-inputs/switch
 ---
 
 import { Playground, PropsTable } from 'docz';
-import { SwitchButton, Colors } from '@class101/ui';
+import { Switch, Colors } from '@class101/ui';
 import { PlayGroundButtonContainer } from '../HelpComponents';
 
-# SwitchButton
+# Switch
 
 ## Props
 
-<PropsTable of={SwitchButton} />
+<PropsTable of={Switch} />
 
 ## Basic usage
 
 ```ts
-import { SwitchButton, Colors } from '@class101/ui';
+import { Switch, Colors } from '@class101/ui';
 ```
 
 <Playground>
   <PlayGroundButtonContainer>
-    <SwitchButton />
-    <SwitchButton size="md" />
-    <SwitchButton color={Colors.blue500} />
+    <Switch />
+    <Switch color={Colors.blue500} />
   </PlayGroundButtonContainer>
 </Playground>
 
@@ -32,7 +31,7 @@ import { SwitchButton, Colors } from '@class101/ui';
 
 <Playground>
   <PlayGroundButtonContainer>
-    <SwitchButton checked />
-    <SwitchButton disabled />
+    <Switch checked />
+    <Switch disabled />
   </PlayGroundButtonContainer>
 </Playground>
