+++ src/FormInput/index.ts
@@ -4,4 +4,4 @@ export { default as Textarea } from './Textarea';
 export { Intent } from './InlineError';
 export { default as InlineError } from './InlineError';
 export { default as TagInput } from './TagInput';
-export { default as SwitchButton } from './SwitchButton';
+export { default as Switch } from './Switch';
